+++ src/tokens_replace_mt.cpp
@@ -13,7 +13,7 @@ using namespace ngrams;
 
 
 Text replace(Text tokens, 
-             int span_max,
+             size_t span_max,
              MapNgrams &map_words){
     
     if(tokens.size() == 0) return {}; // return empty vector for empty text
@@ -39,11 +39,11 @@ struct replace_mt : public Worker{
     
     Texts &input;
     Texts &output;
-    int span_max;
+    size_t span_max;
     MapNgrams &map_words;
     
     // Constructor
-    replace_mt(Texts &input_, Texts &output_, int span_max_, MapNgrams &map_words_):
+    replace_mt(Texts &input_, Texts &output_, size_t span_max_, MapNgrams &map_words_):
               input(input_), output(output_), span_max(span_max_), map_words(map_words_) {}
     
     // parallelFor calles this function with size_t
@@ -76,7 +76,7 @@ List qatd_cpp_tokens_replace(List texts_,
     IntegerVector ids = ids_;
 
     MapNgrams map_words;
-    int span_max = 0;
+    size_t span_max = 0;
     for (size_t g = 0; g < words.size(); g++){
         if(has_na(words[g])) continue;
         Ngram word = words[g];
